+++ block/src/switch/facetwp.js
@@ -1,0 +1,17 @@
+import FacetWP from './compatibility/FacetWP';
+
+window.addEventListener( 'DOMContentLoaded', () => {
+	const viewConfig = window.wtsViewObject || {
+		originalTaxDisplay: 'incl',
+	};
+
+	const wtsCompatibilityObject = window.wtsCompatibilityObject || {
+		baseTaxRate: 0,
+	};
+
+	const facetwp = new FacetWP(
+		viewConfig.originalTaxDisplay,
+		wtsCompatibilityObject.baseTaxRate
+	);
+	facetwp.init();
+} );